Future Fix; Add other way of creating body snake
The while loop that creates the snake body, does so each time the snake
moves,  for the entire snake body array.  This is repetitive.  However,
until the 'apple spawning into the snake' issue has been fully addressed
we can not fully remove the while loop.  As without it, when the apple
spawns into the snake, it ends up removing the snake class and staying
an apple.